+++ nn_dataflow/__init__.py
@@ -42,5 +42,5 @@ from .Scheduling import SchedulingCondition, SchedulingResult, Scheduling
 
 from .NNDataflow import NNDataflow
 
-__version__ = '1.4.0'
+__version__ = '1.5-rc.3'