+++ CHANGELOG.md
@@ -4,6 +4,35 @@ List of major changes and improvements
 ## [Unreleased]
 
 
+## [v1.4 -- v1.5] -- 2017-06-02
+
+### Added
+
+- Software engineering.
+
+  - Add utilities to LoopBlockingScheme class.
+
+
+### Changed
+
+- Software models:
+
+  - Add loop index generator to LoopBlockingScheme class.
+
+  - Integrate verification into LoopBlockingScheme class.
+
+- Software engineering.
+
+  - Lazily evaluate loop blocking stats.
+
+  - Use rich comparison instead of __cmp__.
+
+
+### Fixed
+
+- Output data fetch count.
+
+
 ## [v1.3 -- v1.4] -- 2017-05-18
 
 ### Added